requirement.txt
Django==2.1.7
djangorestframework==3.11.1
django-cors-headers==3.4.0
psycopg2-binary
django-oauth-toolkit